bug 225: fixed nasty 64-bit bug

+++ Appl/CEP/CS1/CS1_Interface/src/Allocator.cc
@@ -64,7 +64,7 @@ void *SparseSetAllocator::allocate(size_t size, unsigned alignment)
 #endif
 
   for (SparseSet<void *>::const_iterator it = freeList.getRanges().begin(); it != freeList.getRanges().end(); it ++) {
-    void *begin = (void *) (((size_t) it->begin + alignment - 1) & ~(alignment - 1));
+    void *begin = (void *) (((size_t) it->begin + alignment - 1) & ~ (size_t) (alignment - 1));
 
     if ((char *) it->end - (char *) begin >= (ptrdiff_t) size) {
       freeList.exclude(begin, (void *) ((char *) begin + size));
